Switch to a continuation-function loop style instead of async.forEach().
The git branch examination is now done serially instead of in parallel.
This is slower but avoids tripping over the too many open files problem
for very large, busy repos. E.g., the one I use this tool with most
often.

Remove the dependency on async since we're not using it any more.

Bring the code into compliance with the Kelp Heavy Weaponry code
formatting standards. As you know, we're hardcore here.

Usage & command line option tweaks. README tweaks to match.
Also, credit where credit is due. So we know who to blame for that
two spaces to indent thing.